+++ biopipen/scripts/delim/SampleInfo.R
@@ -88,7 +88,11 @@ for (name in names(stats)) {
     group <- if (is.null(stat$group)) sym("..group") else sym(stat$group)
     count_on <- paste0("..count.", stat$on)
     if (!is_continuous) {
-        data <- data %>% add_count(!!group, name = count_on)
+        if (!is.null(stat$each)) {
+            data <- data %>% add_count(!!group, !!sym(stat$each), name = count_on)
+        } else {
+            data <- data %>% add_count(!!group, name = count_on)
+        }
     }
 
     if (is.null(stat$devpars)) {
@@ -141,18 +145,19 @@ for (name in names(stats)) {
         } else {
             data <- data %>%
                 distinct(!!group, !!sym(stat$each), .keep_all = TRUE) %>%
+                mutate(!!group := factor(!!group, levels = unique(!!group))) %>%
                 group_by(!!sym(stat$each))
         }
         p <- ggplot(
-            data %>% arrange(!!group),
-            aes(x = "", y = !!sym(count_on), fill = !!group, label = !!sym(count_on))
+            data %>% mutate(.size = sum(!!sym(count_on))),
+            aes(x = sqrt(.size) / 2, width = sqrt(.size), y = !!sym(count_on), fill = !!group, label = !!sym(count_on))
         ) +
-            geom_bar(stat="identity", width=1, color="white", position = position_stack(reverse = TRUE)) +
+            geom_bar(stat="identity", color="white", position = position_fill(reverse = TRUE)) +
             coord_polar("y", start = 0) +
             theme_void() +
             theme(plot.title = element_text(hjust = 0.5)) +
             geom_label_repel(
-                position = position_stack(vjust = 0.5),
+                position = position_fill(reverse = TRUE,vjust = .5),
                 color="#333333",
                 fill="#EEEEEE",
                 size=4